Added task to task processor and other debugging changes
