+++ back-end/blast.py
@@ -24,10 +24,8 @@ import pandas as pd
 import re
 import zipfile
 from sys import getsizeof
-import helper
 from datetime import datetime
 
-response_object = {}
 ROOT = os.path.dirname(os.path.abspath(__file__))
 
 # ------------------------------ MAIN FUNCTIONS ------------------------------
@@ -41,10 +39,12 @@ def find_blast_zip(phage_id):
     Returns:
         A dictionary containing download boolean indicator.
     """
+    response_object = {}
     response_object["blast_downloaded"] = False
     response_object["uploaded"] = True
     response_object["annotated"] = False
     response_object["annotation_in_progress"] = False
+    response_object["blast_input_in_progress"] = False
     if db.session.query(Tasks).filter_by(phage_id=phage_id).filter_by(function="auto_annotate").first() is not None:
         response_object["annotation_in_progress"] = True
     for filename in os.listdir(os.path.join(ROOT, 'users', phage_id, 'uploads')):
@@ -53,9 +53,12 @@ def find_blast_zip(phage_id):
             break
     if (db.session.query(Blast_Results).filter_by(phage_id=phage_id).first() is None):
         response_object["uploaded"] = False
+    if db.session.query(Tasks).filter_by(phage_id=phage_id).filter_by(function="blast_input").first() is not None:
+        response_object["blast_input_in_progress"] = True
     for filename in os.listdir(os.path.join(ROOT, 'users', phage_id)):
         if filename.endswith('.zip'):
             response_object["blast_downloaded"] = True
+            break
     return response_object
 
 def download_blast_input(phage_id):
@@ -71,69 +74,6 @@ def download_blast_input(phage_id):
     f = open(os.path.join(ROOT, 'users', phage_id, f"{phage_id}_blast.zip"), "rb")
 
     return f.read()
-
-def create_blast_input(UPLOAD_FOLDER, phage_id):
-    """Creates fasta file(s) for BLAST input.
-
-    If more than one file is created, then each file should be 30 kb.
-
-    Args:
-        UPLOAD_FOLDER:
-            The path to the directory containing files for the current user.
-        phage_id:
-            The ID of the current user.
-
-    Returns:
-        The Number of blast fasta files created.
-    """
-    fasta_file = helper.get_file_path("fasta", UPLOAD_FOLDER)
-    filename = re.search('(.*/users/.*)/uploads/.*.\w*', fasta_file)
-    genome = SeqIO.read(fasta_file, "fasta").seq
-    output = ""
-    blast_file_count = 1  # keep track of num blast files created
-    out_file = f"{str(filename.group(1))}/{phage_id}_blast_{blast_file_count}.fasta"
-    files_to_zip = [out_file]
-    lefts, rights = get_starts_stops('+', genome)
-    for i in range(len(lefts)):
-        output += f">+, {lefts[i]}-{rights[i]}\n"
-        output += f"{Seq.translate(sequence=helper.get_sequence(genome, '+', lefts[i]-1, rights[i]), table=11)}\n"
-        if getsizeof(output) > 30000:  # only file size to reach 30 kb, else you get a CPU limit from NCBI blast
-            with open(out_file, "w") as f:
-                f.write(output)
-            output = ""
-            blast_file_count += 1
-            out_file = f"{str(filename.group(1))}/{phage_id}_blast_{blast_file_count}.fasta"
-            files_to_zip.append(out_file)
-    lefts, rights = get_starts_stops('-', genome)
-    for i in range(len(lefts)):
-        output += f">-, {lefts[i]}-{rights[i]}\n"
-        left = len(genome) - rights[i]
-        right = len(genome) - lefts[i] + 1
-        output += f"{Seq.translate(sequence=helper.get_sequence(genome, '-', left, right), table=11)}\n"
-        if getsizeof(output) > 30000:  # only file size to reach 30 kb, else you get a CPU limit from NCBI blast
-            with open(out_file, "w") as f:
-                f.write(output)
-            output = ""
-            blast_file_count += 1
-            out_file = f"{str(filename.group(1))}/{phage_id}_blast_{blast_file_count}.fasta"
-            files_to_zip.append(out_file)
-
-    with open(out_file, "w") as f:
-        f.write(output)
-    
-    # zip all out_files together.
-    zip_file = zipfile.ZipFile(f"{str(filename.group(1))}/{phage_id}_blast.zip", 'w', zipfile.ZIP_DEFLATED)
-    for filename in files_to_zip:
-        arcname = filename.rsplit('/', 1)[-1].lower()
-        zip_file.write(filename, arcname)
-    zip_file.close()
-    
-    # delete files that are not in zip folder.
-    for filename in os.listdir(os.path.join(ROOT, 'users', phage_id)):
-        if filename.endswith(".fasta"):
-            os.remove(os.path.join(ROOT, 'users', phage_id, filename))
-
-    return blast_file_count
 
 def dropzone(phage_id, UPLOAD_FOLDER, request):
     """Adds the blast output file to the upload directory if of type json.
@@ -186,6 +126,7 @@ def get_blast_output_names(phage_id, UPLOAD_FOLDER, type_of_call):
     Returns:
         A dictionary containing a list of the blast output file names.
     """
+    response_object = {}
     file_names = []
     file_sizes = []
     file_mods = []
@@ -214,6 +155,7 @@ def get_blast_output_names(phage_id, UPLOAD_FOLDER, type_of_call):
     response_object["in_process"] = False
     response_object["position"] = -1
     response_object["result"] = "not complete"
+    response_object["blast_input_complete"] = False
     task = db.session.query(Tasks).filter_by(phage_id=phage_id).filter_by(function="auto_annotate").first()
     if (task is not None):
         curr_tasks = db.session.query(Tasks).filter_by(complete=False).order_by(Tasks.time)
@@ -229,7 +171,12 @@ def get_blast_output_names(phage_id, UPLOAD_FOLDER, type_of_call):
             response_object["result"] = task.result
             db.session.delete(task)
             db.session.commit()
-
+    task = db.session.query(Tasks).filter_by(phage_id=phage_id).filter_by(function="blast_input").first()
+    if (task and task.complete):
+        response_object["blast_input_complete"] = True
+        response_object["num_files"] = task.result
+        db.session.delete(task)
+        db.session.commit()
     return response_object
 
 def delete_blast_output(phage_id, UPLOAD_FOLDER, file_path):
@@ -246,6 +193,7 @@ def delete_blast_output(phage_id, UPLOAD_FOLDER, file_path):
     Returns:
         A dictionary containing a success message.
     """
+    response_object = {}
     try:
         os.remove(os.path.join(UPLOAD_FOLDER, file_path))
         delete_file = db.session.query(Files).filter_by(phage_id=phage_id).filter_by(name=file_path).first()
@@ -319,13 +267,37 @@ def add_annotation_task(phage_id, UPLOAD_FOLDER):
                 arguments=args,
                 complete=False,
                 result="waiting",
-                time=datetime.now())
+                time=str(datetime.now()))
     try:
         db.session.add(task)
         db.session.commit()
     except:
         return "Error in adding task to queue"
     return "success"
+
+def add_blast_input_task(UPLOAD_FOLDER, phage_id):
+    """Adds task to database to be executed.
+
+    Args:
+        phage_id:
+            The ID of the current user.
+        UPLOAD_FOLDER:
+            The folder containing all the uploaded files.
+    """
+    args = UPLOAD_FOLDER + " " + phage_id
+    task = Tasks(phage_id=phage_id,
+                function="blast_input",
+                arguments=args,
+                complete=False,
+                result="waiting",
+                time='0' + str(datetime.now()))
+    try:
+        db.session.add(task)
+        db.session.commit()
+    except:
+        return "Error in adding task to queue"
+    return "success"
+
 
 def get_num_blast_files(phage_id):
     """Gets the number of Blast input files in the zip folder.
@@ -343,97 +315,3 @@ def get_num_blast_files(phage_id):
                 num_blast_files = len(archive.infolist())
                 return str(num_blast_files)
     return "None"
-
-# ----- BLAST CREATION HELPER FUNCTIONS ------
-def get_stop_options(genome, start, strand):
-    """Finds the next stop codon given a start codon index.
-
-    Args:
-        genome:
-            The genome of the phage.
-        start:
-            The location of the start codon.
-        strand:
-            Complimentary or direct strand.
-
-    Returns:
-        The stop codon index.
-    """
-    
-    bacteria_stop_codons = ["TAG", "TAA", "TGA", "tag", "taa", "tga"]
-    start -= 1
-    gene = ""
-    if (strand != "-"):
-        gene = genome[start:]
-    else:
-        gene = genome.reverse_complement()[start:]
-    for index in range(0, len(gene), 3):
-        codon = gene[index:index + 3]
-        if codon in bacteria_stop_codons:
-            return (start + index + 3)
-
-def get_start_options(genome, maximum, strand, minimum):
-    """Finds all the start codons within a range of DNA indicated by the minimum and maximum parameters.
-
-    Args:
-        genome:
-            The genome of the phage.
-        maximum:
-            The max index to search for start codons.
-        strand:
-            Complimentary or direct strand.
-        minimum:
-            The minimum index to search for start codons.
-    """
-
-    bacteria_start_codons = ["ATG", "GTG", "TTG", "atg", "gtg", "ttg"]
-    start_options = []
-    maximum += 3
-    gene = ""
-    if (strand != "-"):
-        gene = genome[minimum:maximum]
-    else:
-        gene = genome.reverse_complement()[minimum:maximum]
-    for index in range(0, len(gene)):
-        codon = gene[index:index + 3]
-        if codon in bacteria_start_codons:
-            start_options.append(minimum + index + 1)
-    return start_options
-
-def get_starts_stops(strand, genome):
-    """Finds alternative, possible start and stop positions for a given CDS. 
-    
-    Args:
-        cds_id:
-            The ID of the CDS.
-        genome:
-            The genome of the phage.
-
-    Returns:
-        start_options, stop_options: 
-            list of alternative starts and an associated list of alternative stops.
-    """
-
-    genome_length = len(genome)
-    possible_start_options = get_start_options(genome, genome_length - 1, strand, 0)
-
-    start_options = []
-    stop_options = []
-    for start in possible_start_options:
-        stop_options.append(get_stop_options(genome, start, strand))
-        if stop_options[-1] is None:
-            stop_options.pop()
-        elif stop_options[-1] - start < 30:
-            stop_options.pop()
-        else:
-            if strand == '+':
-                start_options.append(start)
-            else:
-                start_options.append(genome_length - start + 1)
-                stop = stop_options.pop()
-                stop_options.append(genome_length - stop + 1)
-
-    if strand == '+':
-        return start_options, stop_options
-    else:
-        return stop_options, start_options
