+++ back-end/settings.py
@@ -10,7 +10,6 @@ import os
 from builtins import FileNotFoundError
 import subprocess
 import models
-import helper
 import shutil
 import pandas as pd
